Renames nested membership class to BaseMembership
This should make it much clearer. The nested Membership is simply an
alias to your membership class. Always. If you set custom_membership
to True in your GalleryMeta, it will not be automatically created.
However, a nested BaseMembership class will be present, so that you
can extend it and not have to set up the relationships by hand.