Lazily evaluate default membership content type
Previously, loading this app before syncdb has been run would result
in an error when the content type filter query was run. Of course,
this means that if this app is installed against a fresh database,
syncdb can't be run, since this app will load (and fail) when you try
to run it!  Thanks to @matthewwithanm for the assist.