Allow membership content_type to be blank
This solves a (probably long-standing) issue where galleries that
accept multiple types of members would require that all 'extra'
items in the inline admin form be filled out when editing the
gallery. The problem was only rarely seen because most of the
time, galleries only accept one type of member.